+++ src/db/models/index.js
@@ -5,7 +5,7 @@ const Sequelize = require('sequelize');
 const basename = path.basename(__filename);
 const env = process.env.NODE_ENV || 'development';
 /* eslint-disable-next-line import/no-dynamic-require */
-const config = require(`${__dirname}/../config/config.js`)[env];
+const config = require(`${__dirname}/../../../config/config.js`)[env];
 const db = {};
 
 let sequelize;
